+++ filter/lib/pragma/filter.rb
@@ -1,0 +1,18 @@
+# frozen_string_literal: true
+
+require "pragma/filter/version"
+
+require 'pragma/filter/base'
+require 'pragma/filter/boolean'
+require 'pragma/filter/equals'
+require 'pragma/filter/ilike'
+require 'pragma/filter/like'
+require 'pragma/filter/scope'
+require 'pragma/filter/where'
+
+module Pragma
+  module Filter
+    class Error < StandardError; end
+    # Your code goes here...
+  end
+end