+++ pragma/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- Pragma::Filter was moved to a separate pragma-filter project
+- Pragma::Macro was moved to a separate pragma-macro project
+- Pragma::Decorator::Error was moved to Pragma::Decorator
+
 ## [2.5.0]
 
 ### Added